% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logistic_regression}
\alias{logistic_regression}
\title{A logistic regression method using Rcpp}
\usage{
logistic_regression(explain, response)
}
\arguments{
\item{explain}{the design matrix}

\item{response}{the response vector y}
}
\value{
the estimated coefficients of variables
}
\description{
A logistic regression method using Rcpp
}
