% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMAP_weights.R
\name{SMAP_weights}
\alias{SMAP_weights}
\title{A SMAP function to estimate the weights for each sub model using R}
\usage{
SMAP_weights(data, p, p_dis, n_D)
}
\arguments{
\item{data}{which data you want to use to build a SMAP model}

\item{p}{the number of variables}

\item{p_dis}{the number of discrete variables}

\item{n_D}{the sample size}
}
\value{
the weights for each sub-model by SMAP and improved SMAP
}
\description{
A SMAP function to estimate the weights for each sub model using R
}
\examples{
\dontrun{
     data<-data(twotypeVehicle)
     w<-SMAP_weights{data,10,3}
}
}
