\name{StatComp22011-package}
\alias{StatComp22011-package}
\docType{package}
\title{
  This package provides several functions to estimate the moving average method weights 
}
\description{
  R version of SMAP and improved-SMAP and Rcpp version of 
}
\author{
Yuanyuan Yan

Maintainer: Yuanyuan Yan <yanyuanyuan@163.com>
}
\keyword{ package }

