% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMAP_weights.R
\docType{data}
\name{twotypeVehicle}
\alias{twotypeVehicle}
\title{A illustration dataset}
\description{
A dataset used to illustrate the performance of SMAP and logistic method.
}
\examples{
\dontrun{
data(twotypeVehicle)
attach(twotypeVehicle)
}
}
